vulkan-server/pyproject.toml
[project]
name="vulkan_server"
version="0.1.0"
description="Vulkan data server"
dependencies = [
    "fastapi[standard]>=0.112.1",
    "pandas>=2.2.2",
    "sqlalchemy>=2.0.32",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.2",
    "dagster-graphql>=1.8.2",
    "pytest>=8.3.2",
    "pytest-httpserver>=1.1.0",
]
requires-python = ">=3.10"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.2",
]

